==> 1. Data Cleaning.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import zipfile as zip
import os

######################################################################################
## FUNCTIONS #########################################################################

def read_zip(filename):
    z = zip.ZipFile("Data/Original data/" + filename, "r")
    df = pd.read_csv(z.open(z.namelist()[0]), header = 0)
    z.close()
    return df

######################################################################################
## INLEZEN ###########################################################################
files = os.listdir('Data/Original data')
li = []   
for filename in files:
    df = read_zip(filename)
    li.append(df)

disney = df = pd.concat(li)

disney = df.copy()

######################################################################################
## MANIPULEREN #######################################################################
## P2
disney = disney[disney["PRICE"] != 0.00]
disney

## P3
disney = disney[disney['EX'] == "T"]
disney

## T1
disney = disney[disney['CORR'] == 0]
disney

## T2: E, F have a problem
disney = disney[disney['COND'].isin(['@', 'F', '6'])]
disney

## T3: gebruik mean
disney['TIME'] = pd.to_datetime(disney['TIME'], format='%H:%M:%S').dt.time
disney = disney.groupby(['DATE', "TIME"]).agg({'SIZE': 'sum', 'PRICE':['median']})
disney.reset_index()

## T4/Q4
deleted_rows = 0
disney_index = disney.reset_index()
prices = disney_index["PRICE"].values
for i, price in enumerate(prices):
    band_price = prices[i-25:i+25]
    med_price = np.median(band_price)
    MAD = np.mean(abs(band_price - med_price))
    if (abs(price - med_price) > 10 * MAD):
        deleted_rows+=1
        disney_index.drop(i, inplace = True)

disney_index

filename = "Disney"
compression_options = dict(method='zip', archive_name=f'{filename}.csv')
disney_index.to_csv(f'Data/{filename}.zip', compression=compression_options, header = True, index = True)
